requirements.txt
alembic==1.4.2
appdirs==1.4.4
attrs==20.1.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-flask==1.0.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
tox==3.19.0
virtualenv==20.0.31
Werkzeug==1.0.1
